@@ -1,6 +1,7 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
 	darkMode: 'media',
-	content: ['./index.html', './src/**/*.{svelte,js,ts,html}'],
+	content: ['./index.html', './src/**/*.{svelte,js,ts,jsx,tsx}'],
 	theme: {
 		extend: {
 			spacing: {
