@@ -1,7 +1,8 @@
+import { mount } from 'svelte';
+import './app.css';
 import App from './App.svelte';
-import './index.css';
 
-const app = new App({
+const app = mount(App, {
 	target: document.getElementById('app')
 });
 
