@@ -1,11 +1,14 @@
 import { defineConfig } from 'vite';
 import { svelte } from '@sveltejs/vite-plugin-svelte';
+import path from 'path';
 
-// https://vitejs.dev/config/
+// https://vite.dev/config/
 export default defineConfig({
 	plugins: [svelte()],
-	build: {
-		outDir: 'S3',
-		emptyOutDir: true
+	resolve: {
+		alias: {
+			$lib: path.resolve('./src/lib'),
+			$components: path.resolve('./src/lib/components')
+		}
 	}
 });
